mqtt-2-bigquery/requirements.txt
#pyarrow
#google-cloud
google-cloud-bigquery
paho-mqtt